+++ DesignToolbox/DesignToolboxSnapshotsTests/__Snapshots__/Components/OrangeInverseTheme/Components/Actions/OrangeInverseTheme+ButtonUITests.swift
@@ -23,25 +23,49 @@ import SwiftUI
 final class OrangeInverseThemeButtonUITests: ButtonUITestsTestCase {
 
     // swiftlint:disable implicitly_unwrapped_optional
-    private var theme: OUDSTheme!
+    private var standardTheme: OUDSTheme!
+    private var tunedTheme: OUDSTheme!
     // swiftlint:enable implicitly_unwrapped_optional
 
     override func setUp() {
-        theme = OrangeInverseTheme()
+        standardTheme = OrangeInverseTheme() // Default tuning do not have rounded corners
+        tunedTheme = OrangeInverseTheme(tuning: Tuning(hasRoundedButtons: true))
     }
 
-    /// Tests all buttons configuration in the `OrangeInverseTheme` with the `light` color schemes.
-    @MainActor func testAllButtonsOrangeInverseThemeLight() {
+    // MARK: Orange Inverse theme - no tuning
+
+    /// Tests all buttons configuration in the `OrangeInverseTheme` with the `light` color schemes
+    /// with default tuning.
+    @MainActor func testAllButtonsOrangeInverseStandardThemeLight() {
         let interfaceStyle = UIUserInterfaceStyle.light
-        testAllButtons(theme: theme, interfaceStyle: interfaceStyle)
-        testAllButtonsOnColoredSurface(theme: theme, interfaceStyle: interfaceStyle)
+        testAllButtons(theme: standardTheme, interfaceStyle: interfaceStyle)
+        testAllButtonsOnColoredSurface(theme: standardTheme, interfaceStyle: interfaceStyle)
     }
 
-    /// Tests all buttons configuration in the `OrangeInverseTheme` with the `dark` color schemes.
-    @MainActor func testAllButtonsOrangeInverseThemeDark() {
+    /// Tests all buttons configuration in the `OrangeInverseTheme` with the `dark` color schemes
+    /// with default tuning.
+    @MainActor func testAllButtonsOrangeInverseStandardThemeDark() {
         let interfaceStyle = UIUserInterfaceStyle.dark
-        testAllButtons(theme: theme, interfaceStyle: interfaceStyle)
-        testAllButtonsOnColoredSurface(theme: theme, interfaceStyle: interfaceStyle)
+        testAllButtons(theme: standardTheme, interfaceStyle: interfaceStyle)
+        testAllButtonsOnColoredSurface(theme: standardTheme, interfaceStyle: interfaceStyle)
+    }
+
+    // MARK: Orange Inverse theme - with tuning
+
+    /// Tests all buttons configuration in the `OrangeInverseTheme` with the `light` color schemes
+    /// with special tuning (rounded corners)
+    @MainActor func testAllButtonsOrangeInverseTunedThemeLight() {
+        let interfaceStyle = UIUserInterfaceStyle.light
+        testAllButtons(theme: tunedTheme, interfaceStyle: interfaceStyle)
+        testAllButtonsOnColoredSurface(theme: tunedTheme, interfaceStyle: interfaceStyle)
+    }
+
+    /// Tests all buttons configuration in the `OrangeInverseTheme` with the `dark` color schemes
+    /// with special tuning (rounded corners)
+    @MainActor func testAllButtonsOrangeInverseTunedThemeDark() {
+        let interfaceStyle = UIUserInterfaceStyle.dark
+        testAllButtons(theme: tunedTheme, interfaceStyle: interfaceStyle)
+        testAllButtonsOnColoredSurface(theme: tunedTheme, interfaceStyle: interfaceStyle)
     }
 }