+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- [Library] Vocalisation of badge with icons (Orange-OpenSource/ouds-ios#956)
+- [Library] Vocalisation of badge with icons (Orange-OpenSource/ouds-ios#957)
 - [Library] Badge component does not have bigger sizes if text sizes is increased (Orange-OpenSource#844)
 
 ## [0.18.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/0.17.0...0.18.0) - 2025-09-05